requirements.txt
sphinx>=4.0
sphinx_rtd_theme
nbsphinx
ipython
jinja2
numpy
numba
pyyaml
xarray
dask
pytest
netCDF4
matplotlib
cftime
scipy
jupyter
cmocean
cartopy
distributed
ncar_jobqueue
PyYAML
intake
nc_time_axis
momlevel @ git+https://github.com/jkrasting/momlevel
xwavelet @ git+https://github.com/jkrasting/xwavelet
